+++ packages/@baldeweg/template-vue/src/i18n.js
@@ -1,14 +1,22 @@
 import { createI18n } from 'vue-i18n'
-import en from '@/../locales/en.json'
-import de from '@/../locales/de.json'
+
+const modules = import.meta.glob('./../locales/*.json')
+
+const getName = (data) => {
+  return data.replace(/..\/locales\/(.*).json/, '$1')
+}
+
+const messages = {}
+for (const path in modules) {
+  modules[path]().then((mod) => {
+    messages[getName(path)] = mod
+  })
+}
 
 export default createI18n({
   legacy: false,
   locale: navigator.language,
   fallbackLocale: 'en',
   allowComposition: true,
-  messages: {
-    en,
-    de,
-  },
+  messages,
 })